Add BiomassStore for biomass observation code
`ObservationStore` is getting unreasonably large. The functionality related to
biomass observations is largely separate from the monitoring observation logic,
so start splitting it off into its own separate class.

This initial change just moves the `insertBiomassDetails` method, but it also
lays some groundwork for later changes.
